requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
boto3
botocore